pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
]
